Replaced uses of `size_t` by `int64_t`
It is simpler to deal with (no risk of wrapping overflow) and large enough.
Unsigned integers bring no guarantee on the positivity of values, they only come
with the risk of negative values wrapping to very large numbers.

The Google C++ Style Guide says:

    In particular, do not use unsigned types to say a number will never be
    negative. Instead, use assertions for this.

Other sources suggest not using unsigned types because they bring more trouble
than benefit:

https://wesmckinney.com/blog/avoid-unsigned-integers
https://www.learncpp.com/cpp-tutorial/unsigned-integers-and-why-to-avoid-them

+++ src/animation_player.cpp
@@ -8,7 +8,7 @@ AnimationPlayer::AnimationPlayer(
 }
 
 uint16_t AnimationPlayer::animatedTileIndexFor(uint16_t tile_index,
-                                               size_t tick) const {
+                                               int64_t tick) const {
   if (!animations_per_tile_index_.contains(tile_index)) {
     return tile_index;
   }
