+++ src/ellipse.cpp
@@ -1,8 +1,8 @@
 #include "ellipse.hpp"
 
-Ellipse::Ellipse(size_t radius_x, size_t radius_y)
+Ellipse::Ellipse(int64_t radius_x, int64_t radius_y)
     : radius_x_{radius_x}, radius_y_{radius_y} {}
 
-size_t Ellipse::radiusX() const { return radius_x_; }
+int64_t Ellipse::radiusX() const { return radius_x_; }
 
-size_t Ellipse::radiusY() const { return radius_y_; }
+int64_t Ellipse::radiusY() const { return radius_y_; }
