+++ src/navigator.hpp
@@ -11,14 +11,14 @@
  */
 class Navigator {
  public:
-  Navigator(Collider const& collider, Level const& level, size_t tiles_size);
+  Navigator(Collider const& collider, Level const& level, int64_t tiles_size);
   Position moveBy(PositionedRectangle const& source_positioned_rectangle,
                   Vector const& desired_displacement) const;
 
  private:
   Collider const& collider_;
   Level const& level_;
-  size_t const tiles_size_;
+  int64_t const tiles_size_;
 
   /**
    * @brief Given a position and a movement (delta) on an axis (representing
@@ -29,11 +29,11 @@ class Navigator {
    * @param delta_value Movement of the object.
    * @param upper_bound Maximum position on the axis (the minimum position is
    * zero).
-   * @return size_t The resulting position of the object on the axis after
+   * @return int64_t The resulting position of the object on the axis after
    * applying the movement, and clamping to the bounds.
    */
-  static size_t clampToBounds(size_t object_position, size_t object_size,
-                              int64_t delta_value, size_t upper_bound);
+  static int64_t clampToBounds(int64_t object_position, int64_t object_size,
+                               int64_t delta_value, int64_t upper_bound);
   // FIXME: Move to Collider?
   bool collidesWithTiles(PositionedRectangle const& positioned_rectangle) const;
   Position findNearestPositionToDestination(
